Add proof of concept shared common.js demo
Add a proof of concept sharing code between the content
script and the background worker via code in common.js.
This is confusing to get working right because the
content script cannot easily use modules. This code
demos it working by running helloWorld in both the
content script and the background worker. Later this
approach will be used to share other code.