Change settings layout to store in v1
Change the layout of the settings to store settings in v1.
This is done to allow version changes to v2, etc., to be
done by making a new layout rather than attempting to
model the old. Also, users of settings have this hidden
from them. They only get the current v* object as their
settings object.