Rough in saving / loading active book sets
Rough in the saving and loading of active books sets.
Currently, it will save to an active book set called
'solid', and it will load from that same active book
set. Also the ability to delete active book sets on
the options page has been added. Forthcoming is the
ability to name active book sets and load an active
book set by name.