Wait for settings to add page enhancements using settings
Rework the page initialization to wait to add page
elements that use settings data until that settings
data has been retrieved. Before, the elements would
be added and then augmented when the data came in.
For example, the elements would be added up front
and then the data list would be populated when the
settings comes back. This simplifies things to just
add such an element in one place, when the settings
data has come back.