Set up transition from settings v1 to v2
Set up the transition of settings v1 to settings v2 with
the addition of the bookLinkTarget value. This value will
indicate the target for new tabs. For now, this is hard
coded to contain _blank. Later users will be able to
specify book instead and this will cause books launches
to go into the same book window by using the book name
as the target.