Fix tab creation position when creating for book links
Fix the position of adding tabs when clicking
on books links. The value is now passed in an
array so that if the called function needed to
create a tab it can use and INCREMENT that value.
If it didn't need to create a tab it just leaves
it alone.