Add a functioning export button to options page
Add a functioning export button to the options page that
allows for the saving of a copy of the current options
to disk. Later, an import button will be added which will
allow the user to restore settings from that file. This
will allow settings to be synced across computers (cause
this not being in the app store does not allow syncing
via regular installation).