Have dashed outline for potentially clickable links
Change the approach to showing a dashed outline for
potentially clickable links, that change to a solid
outlines when the CTRL key is pressed and the special
click hook is put in place. This makes it so the user
can see an indication that something special is attached
to that element.